+++ src/main/java/uk/nhs/tis/trainee/notifications/config/MongoCollectionConfiguration.java
@@ -21,8 +21,6 @@
 
 package uk.nhs.tis.trainee.notifications.config;
 
-import static org.springframework.core.Ordered.HIGHEST_PRECEDENCE;
-
 import jakarta.annotation.PostConstruct;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Configuration;
