+++ src/test/java/uk/nhs/tis/trainee/notifications/service/ProgrammeMembershipServiceTest.java
@@ -972,30 +972,6 @@ class ProgrammeMembershipServiceTest {
   }
 
   @Test
-  void shouldScheduleNotificationIfHistoryStartDateIsDateNotLocalDate() {
-    RecipientInfo recipientInfo = new RecipientInfo("id", MessageType.EMAIL, "test@email.com");
-    Date originalStartDate = Date.from(START_DATE.atStartOfDay(timezone).toInstant()
-        .minus(DEFERRAL_IF_MORE_THAN_DAYS + 1, ChronoUnit.DAYS));
-    LocalDate originalSentAt = LocalDate.now().minusDays(100);
-    TemplateInfo templateInfo = new TemplateInfo(null, null,
-        Map.of(START_DATE_FIELD, originalStartDate));
-
-    History sentNotification = new History(ObjectId.get(),
-        new TisReferenceInfo(PROGRAMME_MEMBERSHIP, TIS_ID),
-        PROGRAMME_CREATED, recipientInfo,
-        templateInfo,
-        Instant.from(originalSentAt.atStartOfDay(timezone)), Instant.MAX,
-        SENT, null, null);
-    Map<NotificationType, History> alreadySent = Map.of(PROGRAMME_CREATED, sentNotification);
-
-    ProgrammeMembership programmeMembership = getDefaultProgrammeMembership();
-
-    boolean shouldSchedule
-        = service.shouldScheduleNotification(PROGRAMME_CREATED, programmeMembership, alreadySent);
-    assertThat("Unexpected should schedule value.", shouldSchedule, is(true));
-  }
-
-  @Test
   void shouldNotScheduleNotificationIfHistoryStartDateCorrupt() {
     RecipientInfo recipientInfo = new RecipientInfo("id", MessageType.EMAIL, "test@email.com");
     LocalDate originalSentAt = LocalDate.now().minusDays(100);
